msm: clock-krait: Work-around retention corner-case To quote Matt Wagantall in his original commit: The CPU clock source MUXes cannot switch while a CPU is in retention, but callers of acpuclock APIs may not known that the CPU is in retention.
Work around this problem be always performing the CPU MUX
switch on the affected CPU using smp_call_function_single()
in cases where cross-calling might otherwise be possible.
The one exception to this is during hotplug code paths,
where cross-calling is expected by design (since the target
CPU is not online and cannot be in retention).

Change-Id: I529e1406f9ce340c9fee30475facd5a17efd991c